+++ src/renderer/renderScene.ts
@@ -849,11 +849,9 @@ export const renderSceneToSvg = (
   {
     offsetX = 0,
     offsetY = 0,
-    exportWithDarkMode = false,
   }: {
     offsetX?: number;
     offsetY?: number;
-    exportWithDarkMode?: boolean;
   } = {},
 ) => {
   if (!svgRoot) {
@@ -870,7 +868,6 @@ export const renderSceneToSvg = (
           files,
           element.x + offsetX,
           element.y + offsetY,
-          exportWithDarkMode,
         );
       } catch (error: any) {
         console.error(error);