+++ src/tests/scene/export.test.ts
@@ -59,9 +59,14 @@ describe("exportToSvg", () => {
       null,
     );
 
-    expect(svgElement.getAttribute("filter")).toMatchInlineSnapshot(
-      `"themeFilter"`,
-    );
+    const css = svgElement.querySelector("style")?.innerHTML;
+    expect(css).not.toBeNull();
+
+    // find the filter value with a regex, to be sure its under the right CSS rule
+    const match = css?.match(/#group-[\w-]+,\s*image\s*\{\s*filter:\s*(.*?);/);
+    expect(match).not.toBeNull();
+
+    expect(match ? match[1] : "").toMatchInlineSnapshot(`"themeFilter"`);
   });
 
   it("with exportPadding", async () => {